complete refactor ocean node handlers + handler registry + gather loggers in common file
