+++ src/components/P2P/handleProtocolCommands.ts
@@ -1,20 +1,12 @@
 import { pipe } from 'it-pipe'
 import { Readable } from 'stream'
 import { toString as uint8ArrayToString } from 'uint8arrays/to-string'
-import StreamConcat from 'stream-concat'
-// export function handleProtocolCommands (sourceStream:any,sinkStream:any) {
 
-import { DownloadHandler } from '../core/downloadHandler.js'
-import { PROTOCOL_COMMANDS, HANDLERS_COMMANDS } from '../../utils/constants.js'
-import { P2PCommandResponse } from '../../@types'
-import { P2P_CONSOLE_LOGGER } from './index.js'
-import { NonceHandler } from '../core/nonceHandler.js'
-import { FeesHandler } from '../core/feesHandler.js'
-import { StatusHandler } from '../core/statusHandler.js'
-import { EncryptHandler } from '../core/encryptHandler.js'
-import { QueryHandler } from '../core/queryHandler.js'
-import { GetDdoHandler, FindDdoHandler } from '../core/ddoHandler.js'
+import { P2P_CONSOLE_LOGGER } from '../../utils/logging/common.js'
+import { Command } from '../../utils/constants.js'
+import { P2PCommandResponse } from '../../@types/OceanNode'
 import { GENERIC_EMOJIS, LOG_LEVELS_STR } from '../../utils/logging/Logger.js'
+import StreamConcat from 'stream-concat'
 
 export class ReadableString extends Readable {
   private sent = false
@@ -39,16 +31,15 @@ export async function handleProtocolCommands(connection: any) {
     true
   )
   P2P_CONSOLE_LOGGER.logMessage('Using ' + connection.connection.remoteAddr, true)
-
   let status = null
-  let task
+  let task: Command
   let statusStream
   let sendStream = null
   /* eslint no-unreachable-loop: ["error", { "ignore": ["ForInStatement", "ForOfStatement"] }] */
   for await (const chunk of connection.stream.source) {
     try {
       const str = uint8ArrayToString(chunk.subarray())
-      task = JSON.parse(str)
+      task = JSON.parse(str) as Command
     } catch (e) {
       status = { httpStatus: 400, error: 'Invalid command' }
       statusStream = new ReadableString(JSON.stringify(status))
@@ -58,77 +49,35 @@ export async function handleProtocolCommands(connection: any) {
     break
   }
   P2P_CONSOLE_LOGGER.logMessage('Performing task: ' + JSON.stringify(task), true)
-
-  const handlersClasses = [
-    DownloadHandler,
-    EncryptHandler,
-    GetDdoHandler,
-    QueryHandler,
-    NonceHandler,
-    StatusHandler,
-    FindDdoHandler,
-    FeesHandler
-  ]
-  const handlers: Record<string, any> = {}
-  handlersClasses.forEach((HandlerClass, index) => {
-    const handlerInstance = new HandlerClass(this)
-    handlers[HANDLERS_COMMANDS[index]] = handlerInstance
-  })
-
+  // we get the handler from the running instance
+  // no need to create a new instance of Handler on every request
+  const handler = this.getCoreHandlers().getHandlerForTask(task)
   let response: P2PCommandResponse = null
-  try {
-    switch (task.command) {
-      case PROTOCOL_COMMANDS.ECHO:
-        status = { httpStatus: 200 }
-        break
-      case PROTOCOL_COMMANDS.DOWNLOAD:
-        response = await handlers[PROTOCOL_COMMANDS.DOWNLOAD].handle(task)
-        break
-      case PROTOCOL_COMMANDS.GET_DDO:
-        response = await handlers[PROTOCOL_COMMANDS.GET_DDO].handle(task)
-        break
-      case PROTOCOL_COMMANDS.QUERY:
-        response = await handlers[PROTOCOL_COMMANDS.QUERY].handle(task)
-        break
-      case PROTOCOL_COMMANDS.ENCRYPT:
-        response = await handlers[PROTOCOL_COMMANDS.ENCRYPT].handle(task)
-        break
-      case PROTOCOL_COMMANDS.NONCE:
-        response = await handlers[PROTOCOL_COMMANDS.NONCE].handle(task)
-        break
-      case PROTOCOL_COMMANDS.STATUS:
-        response = await handlers[PROTOCOL_COMMANDS.STATUS].handle(task)
-        break
-      case PROTOCOL_COMMANDS.FIND_DDO:
-        response = await handlers[PROTOCOL_COMMANDS.FIND_DDO].handle(task)
-        break
-      case PROTOCOL_COMMANDS.GET_FEES:
-        response = await handlers[PROTOCOL_COMMANDS.GET_FEES].handle(task)
-        break
-      default:
-        status = { httpStatus: 501, error: 'Unknown command' }
+  if (handler === null) {
+    status = { httpStatus: 501, error: `No handler found for command: ${task.command}` }
+  } else {
+    try {
+      response = await handler.handle(task)
+      if (response) {
+        // eslint-disable-next-line prefer-destructuring
+        status = response.status
+        sendStream = response.stream
+      }
+      statusStream = new ReadableString(JSON.stringify(status))
+      if (sendStream == null) pipe(statusStream, connection.stream.sink)
+      else {
+        const combinedStream = new StreamConcat([statusStream, sendStream], {
+          highWaterMark: JSON.stringify(status).length // important for reading chunks correctly on sink!
+        })
+        pipe(combinedStream, connection.stream.sink)
+      }
+    } catch (err) {
+      P2P_CONSOLE_LOGGER.logMessageWithEmoji(
+        'handleProtocolCommands Error: ' + err.message,
+        true,
+        GENERIC_EMOJIS.EMOJI_CROSS_MARK,
+        LOG_LEVELS_STR.LEVEL_ERROR
+      )
     }
-
-    if (response) {
-      // eslint-disable-next-line prefer-destructuring
-      status = response.status
-      sendStream = response.stream
-    }
-
-    statusStream = new ReadableString(JSON.stringify(status))
-    if (sendStream == null) pipe(statusStream, connection.stream.sink)
-    else {
-      const combinedStream = new StreamConcat([statusStream, sendStream], {
-        highWaterMark: JSON.stringify(status).length // important for reading chunks correctly on sink!
-      })
-      pipe(combinedStream, connection.stream.sink)
-    }
-  } catch (err) {
-    P2P_CONSOLE_LOGGER.logMessageWithEmoji(
-      'handleProtocolCommands Error: ' + err.message,
-      true,
-      GENERIC_EMOJIS.EMOJI_CROSS_MARK,
-      LOG_LEVELS_STR.LEVEL_ERROR
-    )
   }
 }
