+++ src/components/core/nonceHandler.ts
@@ -2,12 +2,9 @@ import { Handler } from './handler.js'
 import { P2PCommandResponse } from '../../@types/OceanNode.js'
 import { NonceCommand } from '../../utils/constants.js'
 import { NonceDatabase } from '../database/index.js'
-import {
-  getDefaultErrorResponse,
-  getDefaultResponse,
-  DB_CONSOLE_LOGGER
-} from './utils/nonceHandler.js'
+import { getDefaultErrorResponse, getDefaultResponse } from './utils/nonceHandler.js'
 import { GENERIC_EMOJIS, LOG_LEVELS_STR } from '../../utils/logging/Logger.js'
+import { DB_CONSOLE_LOGGER } from '../../utils/logging/common.js'
 
 export class NonceHandler extends Handler {
   async handle(task: NonceCommand): Promise<P2PCommandResponse> {
