+++ src/components/core/utils/findDdoHandler.ts
@@ -1,8 +1,9 @@
-import { OceanP2P, CACHE_TTL, P2P_CONSOLE_LOGGER } from '../../P2P/index.js'
+import { OceanP2P, CACHE_TTL } from '../../P2P/index.js'
 import { FindDDOCommand } from '../../../utils/constants.js'
 import { LOG_LEVELS_STR } from '../../../utils/logging/Logger.js'
 import { FindDDOResponse } from '../../../@types/index.js'
 import { Service } from '../../../@types/DDO/Service.js'
+import { P2P_CONSOLE_LOGGER } from '../../../utils/logging/common.js'
 
 /**
  * Check if the specified ddo is cached and if the cached version is recent enough
