+++ src/components/httpRoutes/index.ts
@@ -7,23 +7,9 @@ import { providerRoutes } from './provider.js'
 import { aquariusRoutes } from './aquarius.js'
 import { rootEndpointRoutes } from './rootEndpoint.js'
 import { downloadRoute } from './download.js'
-import {
-  CustomNodeLogger,
-  LOGGER_MODULE_NAMES,
-  LOG_LEVELS_STR,
-  defaultConsoleTransport,
-  getCustomLoggerForModule
-} from '../../utils/logging/Logger.js'
-
 export * from './getOceanPeers.js'
 
 export const httpRoutes = express.Router()
-// use this logger instance on all HTTP related stuff
-export const HTTP_LOGGER: CustomNodeLogger = getCustomLoggerForModule(
-  LOGGER_MODULE_NAMES.HTTP,
-  LOG_LEVELS_STR.LEVEL_INFO,
-  defaultConsoleTransport
-)
 
 httpRoutes.use(getOceanPeersRoute)
 httpRoutes.use(getP2PPeersRoute)
